Move user ID validation from the command file to the relevant logic file for adding completers Fixes #299
